backend/requirements.txt
Flask
flask_cors
fastapi
uvicorn
PyGithub
scikit-learn
numpy
pydantic
python-dotenv
requests
langchain
langchain-core
langchain-community
langchain-groq
tqdm
regex
httpx
typing-extensions
textwrap3
waitress
gunicorn
langchain-google-genai 
faiss-cpu
PyGithub